### 2024-11-12 #7464 @editorialbot
[PRE REVIEW]: Analytic continuation component of the GreenX library: robust Padé approximants with symmetry constraints
**Submitting author:** @moritzleucke (<a href="http://orcid.org/0009-0003-4381-0935">Moritz Leucke</a>)
**Repository:** https://github.com/nomad-coe/greenX
**Branch with paper.md** (empty if default branch): joss_analyticcontinuation
**Version:** v1.0
