### 2024-11-12 #7464 @arfon
Note from the submitting author copied here:

> This submission presents the Analytic Continuation component, an independent part of the GreenX software suite. The scope of this publication is the analytic continuation of complex-valued functions using Padé approximants, with a Fortran API. The primary application of this component is in electronic structure methods, though it is broadly applicable across fields requiring analytic continuation. A previous JOSS publication covered the TimeFrequency component, which provides minimax Time-Frequency Grids for the random phase approximation and GW methods. This new submission about the Analytic Continuation component expands the functionality of GreenX without overlap with the previously published work.
