### 2024-11-12 #7464 @moritzleucke
Dear @arfon,

Thank you for reaching out. The new addition to the codebase is the entire folder "GX-AnalyticContinuation" (https://github.com/nomad-coe/greenX/tree/main/GX-AnalyticContinuation) within the GreenX repository. This folder contains the analytic continuation component, which is the focus of the newly submitted paper.

It has no overlap with the time-frequency component in the folder "GX-TimeFrequency", that was published in a previous joss paper (https://doi.org/10.21105/joss.05570).

Best,
Moritz Leucke
