### 2024-12-04 #7464 @Kevin-Mattheus-Moerman
- [ ] Please study the above reference check :point_up: and see if you can address any of the reported potential DOI issues. You can add/amend DOI entries in your `.bib` file, and call `@editorialbot check references` here to check them again. If not all references are used please also clean the .bib file to only contain the relevant references.
