### 2024-12-04 7e93f2a
add DOIs to references
