+++ JOSS/analyticcontinuation/refs.bib
@@ -6,6 +6,7 @@
   number={12},
   pages={5665--5687},
   year={2015},
+  doi = {10.1021/acs.jctc.5b00453},
   publisher={ACS Publications}
 }
 
@@ -17,6 +18,7 @@
   number={16},
   pages={165109},
   year={2016},
+  doi = {10.1103/PhysRevB.94.165109},
   publisher={APS}
 }
 @article{wilhelm2018toward,
@@ -27,6 +29,7 @@
   number={2},
   pages={306--312},
   year={2018},
+  doi = {10.1021/acs.jpclett.7b02740.s001},
   publisher={ACS Publications}
 }
 @article{duchemin2020robust,
@@ -37,6 +40,7 @@
   number={3},
   pages={1742--1756},
   year={2020},
+  doi = {10.1021/acs.jctc.9b01235.s001},
   publisher={ACS Publications}
 }
 @article{friedrich2019tetrahedron,
@@ -47,6 +51,7 @@
   number={7},
   pages={075142},
   year={2019},
+  doi = {10.1103/PhysRevB.100.075142},
   publisher={APS}
 }
 @article{voora2020molecular,
@@ -57,6 +62,7 @@
   number={1},
   pages={433--439},
   year={2020},
+  doi = {10.1021/acs.jpclett.0c03362.s001},
   publisher={ACS Publications}
 }
 @article{samal2022modeling,
@@ -67,6 +73,7 @@
   number={12},
   pages={7272--7285},
   year={2022},
+  doi = {10.1021/acs.jctc.2c00647.s002},
   publisher={ACS Publications}
 }
 @article{springer1998first,
@@ -77,6 +84,7 @@
   number={11},
   pages={2389},
   year={1998},
+  doi = {10.1103/PhysRevLett.80.2389},
   publisher={APS}
 }
 @article{ren2012resolution,
@@ -87,6 +95,7 @@
   number={5},
   pages={053020},
   year={2012},
+  doi = {10.1088/1367-2630/14/5/053020},
   publisher={IOP Publishing}
 }
 @article{gonze2009abinit,
@@ -97,6 +106,7 @@
   number={12},
   pages={2582--2615},
   year={2009},
+  doi = {10.1016/j.cpc.2009.07.007},
   publisher={Elsevier}
 }
 @article{wilhelm2016gw,
@@ -107,6 +117,7 @@
   number={8},
   pages={3623--3635},
   year={2016},
+  doi = {10.1021/acs.jctc.6b00380.s001},
   publisher={ACS Publications}
 }
 @article{wilhelm2017periodic,
@@ -117,6 +128,7 @@
   number={23},
   pages={235123},
   year={2017},
+  doi = {10.1103/PhysRevB.95.235123},
   publisher={APS}
 }
 @article{wilhelm2021low,
@@ -127,16 +139,18 @@
   number={3},
   pages={1662--1677},
   year={2021},
+  doi = {10.1021/acs.jctc.0c01282.s001},
   publisher={ACS Publications}
 }
 @article{forster2020low,
-  title={Low-order scaling g 0 w 0 by pair atomic density fitting},
+  title={Low-order scaling G0W0 by pair atomic density fitting},
   author={Förster, A. and Visscher, L.},
   journal={Journal of chemical theory and computation},
   volume={16},
   number={12},
   pages={7381--7399},
   year={2020},
+  doi = {10.1021/acs.jctc.0c00693},
   publisher={ACS Publications}
 }
 @article{forster2021low,
@@ -146,6 +160,7 @@
   volume={9},
   pages={736591},
   year={2021},
+  doi = {10.3389/fchem.2021.736591},
   publisher={Frontiers Media SA}
 }
 @article{havu2009efficient,
@@ -156,6 +171,7 @@
[…]
