### 2024-12-04 d6271d6
add DOIs to references
