+++ JOSS/analyticcontinuation/refs.bib
@@ -309,7 +309,6 @@ note ={PMID: 37566917},
   author={Thiele, T. N.},
   year={1909},
   publisher={Leipzig B.G. Teubner},
-  doi = {10.2307/2345397},
   url={https://archive.org/search.php?query=external-identifier%3A%22urn%3Aoclc%3Arecord%3A1047515853%22}
 }
 
