+++ JOSS/analyticcontinuation/refs.bib
@@ -574,6 +574,7 @@ year = {2015}
 
 @misc{Montmorency,
   author = {Montmorency},
+  title = {Pade-py},
   publisher = {GitHub},
   year = {2014},
   journal = {GitHub repository},
@@ -582,7 +583,7 @@ year = {2015}
 
 @misc{jjgoings,
   author = {jjgoings},
-  title = {Pade-py},
+  title = {Pade: Fourier transform via Pade approximants},
   publisher = {GitHub},
   year = {2017},
   journal = {GitHub repository},
@@ -591,7 +592,7 @@ year = {2015}
 
 @misc{bennosski,
   author = {bennosski},
-  title = {Pade-py},
+  title = {pade},
   publisher = {GitHub},
   year = {2019},
   journal = {GitHub repository},
