### 2024-12-04 #7464 @moritzleucke
Dear @Kevin-Mattheus-Moerman,

Thanks a lot for the comments on our submission.

We added a `CONTRIBUTING.md`and `CODE_OF_CONDUCT.md` to the [Repository](https://github.com/nomad-coe/greenX).

Additionally, the bibtex file is now updated and the remaining skiped DOIs are GitHub Repositories ("Pade_approximants", "Pade-py", "Pade: Fourier transform via Pade approximants", "pade", "RobustPade.jl") and Books ("GNU MP 6.3 Multiple precision arithmetic library", "The Calculus Of Finite Differences", "Essentials of Padé approximants", "Numerical recipes 3rd edition", "Interpolationsrechnung") with no DOIs.

The suggested DOI for the book "Interpolationsrechnung" is unfortunately only a review for the book and not the DOI of the Book itself.

Best,
Moritz Leucke
