### 2024-12-12 #7464 @zhubonan
@Kevin-Mattheus-Moerman this feels a bit outside my domain, but I can take it if there is no other suitable editor around.
