### 2025-02-19 #7464 @moritzleucke
Hi @HaoZeke,

Thanks a lot for initiating the review process. If it is difficult to find a reviewer, we suggest also considering the following potential reviewers, who have previously reviewed a JOSS publication for another component of GreenX:

| Name | GitHub Username |
|------|----------------|
| Xu He | mailhexu |
| Dario A. Leon | DarioALeonValido |

Additionally, here are other potential reviewers who are not currently associated with JOSS:

| Name | GitHub Username | Email |
|------|----------------|-------------------------|
| Fabien Bruneval | bruneval | [fabien.bruneval@cea.fr](mailto:fabien.bruneval@cea.fr) |
| Sergei Iskakov | iskakoff | [siskakov@umich.edu](mailto:siskakov@umich.edu) |
| Daniel Mejia-Rodriguez | dmejiar | [daniel.mejia@pnnl.gov](mailto:daniel.mejia@pnnl.gov) |
| Kenneth Lopata | klopata | [klopata@lsu.edu](mailto:klopata@lsu.edu) |

Let us know if we can assist further!
