### 2025-02-19 2e61fa8
GX-minimax: add regularization as option
