### 2025-02-19 73ca1e9
fix edge case of pade where all function values are 0 + add test for that case
