### 2025-02-19 3db8392
update deprecated actions/upload-artifact version in paper action pipeline
