### 2025-02-25 55ef641
add doc for regularization
