### 2025-02-25 45e3e25
changing number of components build by default
