+++ README.md
@@ -53,37 +53,37 @@ configure with:
 cmake ../ -DBUILD_SHARED_LIBS=OFF
 ```
 
-Specific shared libraries can be disable, e.g to disable the Projector-Augmented Wave (PAW) component of GreenX, run CMake configure with: 
+Specific shared libraries can be disabled or enabled, e.g to enable the Projector-Augmented Wave (PAW) component of GreenX, run CMake configure with: 
 
 ```bash
-cmake ../ -DPAW_COMPONENT=OFF
+cmake ../ -DPAW_COMPONENT=ON
 ```
 
 Available options to disable one or more components
 
-| Component                              | CMake configure                    |
-|----------------------------------------|------------------------------------|
-| Analytical Continuation component      | `cmake .. -DAC_COMPONENT=OFF`      |
-| Minimax Time-Frequency grids component | `cmake .. -DMINIMAX_COMPONENT=OFF` |
-| Localized Basis component              | `cmake .. -DLBASIS_COMPONENT=OFF`  |
-| Projector-Augmented Wave component     | `cmake .. -DPAW_COMPONENT=OFF`     |
+| Component                              | CMake configure                    | Default |
+|----------------------------------------|------------------------------------|---------|
+| Analytical Continuation component      | `-DAC_COMPONENT`      | `ON`  |
+| Minimax Time-Frequency grids component | `-DMINIMAX_COMPONENT` | `ON`  |
+| Localized Basis component              | `-DLBASIS_COMPONENT`  | `OFF` |
+| Projector-Augmented Wave component     | `-DPAW_COMPONENT`     | `OFF` |
 
-GreenX uses GNU Multiple Precision Arithmetic Library by default in the Analytical Continuation component, you can disable it without any harm by runing CMake configure with:
+GreenX uses GNU Multiple Precision Arithmetic Library by default in the Analytical Continuation component, you can disable it without any harm by running CMake configure with:
 
 ```bash
 cmake ../ -DENABLE_GNU_GMP=OFF
 ```
 
-GreenX uses submodules and they are built by default. You can obtained them by executing:
+To build GreenX with submodules (they are turned off by default), one can configure with:
+
+```bash
+cmake ../ -DCOMPILE_SUBMODULES=ON
+```
+
+You can obtained them by executing:
 
 ```bash
 git submodule update --init --recursive
-```
-
-To build GreenX without the submodules, one can configure with:
-
-```bash
-cmake ../ -DCOMPILE_SUBMODULES=OFF
 ```
 
 If all requirements are found, build and install the project:
@@ -96,14 +96,15 @@ make install
 Minimal example to build and install only the Time-Frequency component of GreenX:
 
  ```bash
-cmake .. -DAC_COMPONENT=OFF -DLBASIS_COMPONENT=OFF -DPAW_COMPONENT=OFF -DCOMPILE_SUBMODULES=OFF
+cmake .. -DAC_COMPONENT=OFF 
 make -j
 make install
  ```
 
 
-## Running the Tests
+## Regression Tests
 
+### Running Regression Tests
 GreenX uses pytest as its regression testing framework, in conjunction with 
 the custom python module `pygreenx`. First, one must ensure that `pygreenx`
 is installed. From the GreenX root directory:
@@ -130,43 +131,67 @@ cd build
 ctest
  ```
 
-## Building Documentation
+#### Adding Regression Tests 
 
-GreenX is documented using Doxygen, and documentation support is disabled by
-default. To enable CMake looking for Doxygen, configure with:
 
-```bash
-cmake ../ -DENABLE_GREENX_DOCS=ON
-```
[…]
