### 2025-02-25 1fba977
enable all components explicitely in the pipeline
