### 2025-02-28 #7859 @editorialbot
[REVIEW]: Analytic continuation component of the GreenX library: robust Padé approximants with symmetry constraints
**Submitting author:** @moritzleucke (<a href="http://orcid.org/0009-0003-4381-0935">Moritz Leucke</a>)
**Repository:** https://github.com/nomad-coe/greenX
**Branch with paper.md** (empty if default branch): joss_analyticcontinuation
**Version:** v2.2
