### 2025-03-09 507d095
add generation of greenXConfig.cmake enabling usage with find_package
