+++ GX-common/CMakeLists.txt
@@ -1,6 +1,7 @@
 # CMakeLists.txt for GreenX Common Modules, required by all other libs in
 # the package
 
+
 add_library(GXCommon "")
 
 set_target_properties(GXCommon
@@ -12,7 +13,12 @@ set_target_properties(GXCommon
         LIBRARY_OUTPUT_DIRECTORY ${CMAKE_Fortran_LIB_DIRECTORY}
         )
 
-target_include_directories(GXCommon PUBLIC src/)
+#target_include_directories(GXCommon PUBLIC src/)
+target_include_directories(GXCommon
+    PUBLIC
+        $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/src>  # used when building in the source tree
+        $<INSTALL_INTERFACE:include>                        # used after 'make install'
+)
 
 target_sources(GXCommon PRIVATE
         src/kinds.f90
@@ -26,8 +32,10 @@ target_sources(GXCommon PRIVATE
 # Destination relative to ${CMAKE_INSTALL_PREFIX}, defined in the top-level
 # CMakeLists.txt
 install(TARGETS GXCommon
-        ARCHIVE DESTINATION lib
-        LIBRARY DESTINATION lib)
+    EXPORT greenXTargets
+    ARCHIVE DESTINATION lib
+    LIBRARY DESTINATION lib
+)
 
 # Install modules
 # Destination relative to ${CMAKE_INSTALL_PREFIX}, defined in the top-level
