### 2025-03-12 dc626cb
correct path to module path (INSTALL_INTERFACE)
