+++ GX-LocalizedBasis/src/localized_basis_types.f90
@@ -29,7 +29,32 @@ module localized_basis_types
       integer                                    :: n_species
       character, dimension(:), allocatable       :: species
       real(kind=8), dimension(:,:), allocatable  :: coords
-   end type   
+   end type
+
+   type minimax_types
+      integer                                   :: n_points ! number of points
+      real(kind=8), dimension(:,:), allocatable :: cos_tf   ! transformation
+      real(kind=8), dimension(:),   allocatable :: omega    ! frequency points
+      real(kind=8), dimension(:),   allocatable :: tau      ! time points
+      real(kind=8), dimension(:),   allocatable :: weights  ! rpa weights
+   end type minimax_types
+
+   type kohn_sham_types
+      real(kind=8), dimension(:,:,:), allocatable :: wave         ! KS wave function
+      real(kind=8), dimension(:,:),   allocatable :: occupied     ! occ wave function
+      real(kind=8), dimension(:,:),   allocatable :: virtual      ! virt wave function
+      real(kind=8), dimension(:,:,:), allocatable :: eigenvector  ! KS eigenvectors
+      real(kind=8), dimension(:,:),   allocatable :: eigenvalue   ! KS eigenvalues
+   end type kohn_sham_types
+
+   type polarizabily_types
+      real(kind=8), dimension(:,:),   allocatable :: chi_kk         ! Polarizability (r_k)
+      real(kind=8), dimension(:,:),   allocatable :: green_forward  ! Green function (tau)
+      real(kind=8), dimension(:,:),   allocatable :: green_backward ! Green function (-tau)
+      real(kind=8), dimension(:,:),   allocatable :: pi_pq_tau      ! Polarizability (tau)
+      real(kind=8), dimension(:,:,:), allocatable :: pi_pq_omega    ! Polarizability (omega)
+      real(kind=8), dimension(:,:),   allocatable :: rirs_coeff     ! RI-RS Coeff M_{mu k}  
+   end type polarizabily_types
 
    ! Global types
 
@@ -46,5 +71,18 @@ module localized_basis_types
      type(grids_types)                           :: grids
      type(species_types)                         :: species
    end type separable_ri_types
+
+
+   type screened_coulomb_types
+
+      real(kind=8)                                :: error
+      
+      type(minimax_types)                         :: minimax
+      type(kohn_sham_types)                       :: ks
+      type(polarizabily_types)                    :: pi_pq
+
+      real(kind=8), dimension(:,:,:), allocatable :: matrix 
+
+   end type screened_coulomb_types 
 
 end module localized_basis_types
