### 2025-03-27 f6acd42
Added internal calling of minimax grids
