+++ GX-LocalizedBasis/src/localized_basis_environments.f90
@@ -8,13 +8,10 @@
 module localized_basis_environments
 
    use kinds,                 only: dp
-   use error_handling,    only: register_exc   
+   use error_handling,        only: register_exc   
    use localized_basis_types, only: separable_ri_types, &
-                                    screened_coulomb_types
+                                    polarizability_types
    use lapack_interfaces,     only: dgemm, dsyevx
-   use gx_minimax,            only: gx_minimax_grid
-   
-
 
    implicit none
 
@@ -28,13 +25,12 @@ module localized_basis_environments
 
    type(separable_ri_types) :: ri_rs
 
-   ri_rs%basis%n_basbas=303
-   ri_rs%basis%n_basis = 58 
-   ri_rs%basis%n_loc_basbas=303
-   ri_rs%basis%n_basis_pairs=1711
-   ri_rs%n_points = 638
-
-   ri_rs%error = 0.d0
+   ri_rs%basis%n_basbas      =303
+   ri_rs%basis%n_basis       = 58 
+   ri_rs%basis%n_loc_basbas  = 303
+   ri_rs%basis%n_basis_pairs = 1711
+   ri_rs%n_points            = 638
+   ri_rs%error               = 0.d0
 
    if (.not.allocated(ri_rs%ovlp2fn)) then
       allocate(ri_rs%ovlp2fn(ri_rs%basis%n_basis_pairs,ri_rs%n_points))
@@ -50,49 +46,188 @@ module localized_basis_environments
    
    end subroutine initialization
 
-! *********************************************************************
-!> brief allocate the minimax time-frequency grids 
-!  o ri_rs --  separable resolution-of-the identity environment
-! *********************************************************************
-   subroutine initialize_minimax_grids(w_pq)
+  ! **********************************************************************
+  !> \brief Initialize kohn sham array
+  !  param[inout] pi_pq:  polarizability environment
+  ! **********************************************************************   
+   subroutine initialize_kohn_sham (pi_pq)
 
-     type(screened_coulomb_types)               :: w_pq
+     type(polarizability_types)               :: pi_pq
 
-     ! Hard coded (change later) 
-     w_pq%minimax%n_points = 6
-
-     if (.not.allocated(w_pq%minimax%cos_tf)) then
-        allocate(w_pq%minimax%cos_tf(w_pq%minimax%n_points, w_pq%minimax%n_points))
+     if (.not.allocated(pi_pq%ks%eigenvalues)) then
+        allocate(pi_pq%ks%eigenvalues(pi_pq%ks%n_states, 1))
      end if
 
-     if (.not.allocated(w_pq%minimax%omega)) then
-        allocate(w_pq%minimax%omega(w_pq%minimax%n_points))
+     if (.not.allocated(pi_pq%ks%eigenvectors)) then
+        allocate(pi_pq%ks%eigenvectors(pi_pq%ks%n_basis, pi_pq%ks%n_states, 1))
      end if
 
-     if (.not.allocated(w_pq%minimax%tau)) then
-        allocate(w_pq%minimax%tau(w_pq%minimax%n_points))
+     if (.not.allocated(pi_pq%ks%wave)) then
+        allocate(pi_pq%ks%wave(pi_pq%ks%n_basis, pi_pq%ri_rs%n_points, 1))
      end if
 
-     if (.not.allocated(w_pq%minimax%weights)) then
-        allocate(w_pq%minimax%weights(w_pq%minimax%n_points))
+   end subroutine initialize_kohn_sham        
+
+  ! *********************************************************************
+  !> brief allocate the minimax time-frequency grids 
+  !  param[inout] pi_pq:  polarizability environment
[…]
