+++ GX-LocalizedBasis/src/polarizability.f90
@@ -1,0 +1,332 @@
+module polarizability 
+
+  use kinds,                        only: dp
+  use lapack_interfaces,            only: dgemm
+  use localized_basis_environments
+  use separable_ri,                 only: get_rirs_coefficients
+  use gx_minimax,            only: gx_minimax_grid
+  
+
+  implicit none
+
+  contains
+
+  !> brief Compute the irreducible polarizability [PI(iw)]_PQ using 
+  !>       the separable resolution of the identity method   
+  !! @param[in] n_basis: Number of orbital basis (dimension 1 of eigenvectors array) 
+  !! @param[in] n_basis_pairs: Number of orbital basis pairs (dimension 1 of ovlpXfn array)
+  !! @param[in] n_loc_basbas: Number of auxiliary basis fuctions (dimension 2 of ovlp3fn array)
+  !! @param[in] n_rk_points: Number of real-space grid points (dimension 2 of ovlp2fn and
+  !!                         wave arrays )
+  !! @param[in] n_states: Number of Kohn-Sham states (dimension 1 of eingenvalues and wave
+  !!                      arrays  and dimension 2 of the eigenvectors array) 
+  !! param[in] eigenvalues: real array, the Kohn-Shan eigenvalues from the SCF calculation
+  !! param[in] eigenvectors: real array, the Kohn-Sham eigenvectos form the SCF calculation
+  !! param[in] ovlp_2fn: real array, the product of two NAO basis functions
+  !! param[in] ovlp_3fn: real array, the three-center overlap integral over two
+  !!                       NAO basis functions and one auxiliary basis function.
+  !! param[in] wave: real array, the Kohn-Sham wave function in the real-space grid
+  !! param[out] error: real number, maximum error between Coulomb and real-space resolution of
+  !!                   the identity fitting coefficients.          
+  subroutine gx_rirs_polarizability(n_basis, n_basis_pairs, n_basbas, n_rk_points, n_states, &
+                                     eigenvalues, eigenvectors, ovlp2fn, ovlp3fn, wave, error)
+
+   integer                                              :: n_basis, n_basis_pairs, & 
+                                                           n_basbas, n_states, n_rk_points 
+
+   real(kind=dp)                                        :: error
+   real(kind=dp), dimension(n_states)                   :: eigenvalues
+   real(kind=dp), dimension(n_basis, n_states)          :: eigenvectors
+   real(kind=dp), dimension(n_basis_pairs, n_rk_points) :: ovlp2fn
+   real(kind=dp), dimension(n_basis_pairs, n_basbas)    :: ovlp3fn
+   real(kind=dp), dimension(n_states, n_rk_points)      :: wave
+                             
+
+    type(polarizability_types) :: pi_pq
+
+    ! Get the separable RI coefficients 
+    call get_rirs_coefficients(pi_pq%ri_rs, n_basis_pairs, n_basbas, n_rk_points, &
+                               ovlp2fn, ovlp3fn) 
+
+    ! Get the Kohn-Sham wave function in the real-space grid
+    call get_ks_wave(pi_pq, n_basis, n_states, n_rk_points, eigenvalues, & 
+                     eigenvectors, wave)
+                       
+    ! Get minimax time-frequency grid and tranformation matrix from GreenX
+    call get_minimax_grids(pi_pq)
+                       
[…]
