### 2025-03-31 1f5bb82
Added w_engine to obtain the screened coulomb interaction
