+++ GX-LocalizedBasis/src/polarizability.f90
@@ -1,10 +1,22 @@
+! **************************************************************************************************
+!  Copyright (C) 2020-2024 GreenX library
+!  This file is distributed under the terms of the APACHE2 License.
+!
+! **************************************************************************************************
+!> \brief This module contains the subroutines for the localized basis set component of the library
+! ***************************************************************************************************
 module polarizability 
 
   use kinds,                        only: dp
+  use error_handling,               only: register_exc  
   use lapack_interfaces,            only: dgemm
-  use localized_basis_environments
+  use localized_basis_types,        only: polarizability_types
+  use localized_basis_environments, only: initialize_kohn_sham, &
+                                          initialize_minimax_grids, &
+                                          initialize_polarizability, &
+                                          deallocate_polarizability
   use separable_ri,                 only: get_rirs_coefficients
-  use gx_minimax,            only: gx_minimax_grid
+  use gx_minimax,                   only: gx_minimax_grid
   
 
   implicit none
@@ -29,17 +41,17 @@ module polarizability
   !! param[out] error: real number, maximum error between Coulomb and real-space resolution of
   !!                   the identity fitting coefficients.          
   subroutine gx_rirs_polarizability(n_basis, n_basis_pairs, n_basbas, n_rk_points, n_states, &
-                                     eigenvalues, eigenvectors, ovlp2fn, ovlp3fn, wave, error)
+                                    eigenvalues, eigenvectors, ovlp2fn, ovlp3fn, wave, error)
 
-   integer                                              :: n_basis, n_basis_pairs, & 
-                                                           n_basbas, n_states, n_rk_points 
+    integer                                              :: n_basis, n_basis_pairs, & 
+                                                            n_basbas, n_states, n_rk_points 
 
-   real(kind=dp)                                        :: error
-   real(kind=dp), dimension(n_states)                   :: eigenvalues
-   real(kind=dp), dimension(n_basis, n_states)          :: eigenvectors
-   real(kind=dp), dimension(n_basis_pairs, n_rk_points) :: ovlp2fn
-   real(kind=dp), dimension(n_basis_pairs, n_basbas)    :: ovlp3fn
-   real(kind=dp), dimension(n_states, n_rk_points)      :: wave
+    real(kind=dp)                                        :: error
+    real(kind=dp), dimension(n_states)                   :: eigenvalues
+    real(kind=dp), dimension(n_basis, n_states)          :: eigenvectors
+    real(kind=dp), dimension(n_basis_pairs, n_rk_points) :: ovlp2fn
+    real(kind=dp), dimension(n_basis_pairs, n_basbas)    :: ovlp3fn
+    real(kind=dp), dimension(n_states, n_rk_points)      :: wave
                              
 
     type(polarizability_types) :: pi_pq
@@ -62,11 +74,71 @@ module polarizability
     call evaluate_polarizability(pi_pq)    
     
     ! Deallocate working arrays
-    call deallocate_polarizability(pi_pq, .true.)
+    call deallocate_polarizability(pi_pq)
 
-    error=0.d0
+    error=0.0_dp
 
[…]

[…]
