### 2025-04-10 nomad-coe/greenX#156 @mailhexu
GX-AC Tests fail when not build with GMP
Hello,

I compiled GreenX-Analytic-continuition with and without GMP. Then I run ctest.
All the test passed when it is compiled with GMP, and those fails when it is compiled without GMP.

I understand from the following paragraph that this is due to the numerical instability.

```
Padé approximants are known to be [numerical unstable](https://doi.org/10.1093/imamat/25.3.267). The GX-AC component uses two strategies to numerically stabilize the interpolation. First, it incorporates a [greedy algorithm for Thiele Padé approximants](https://pubs.acs.org/doi/full/10.1021/acs.jctc.3c00555) that minimizes the numerical error by reordering the reference points. A validation of the greedy algorithm can be found in this [reference](https://pubs.acs.org/doi/full/10.1021/acs.jctc.3c00555). Additionally, it is possible to use the component with a higher internal numerical floating point precision. This helps reducing the numerical noise caused by [catastrophic cancellation](https://doi.org/10.1145/103162.103163). Catastrophic cancellation occurs when rounding errors are amplified through the subtraction of rounded numbers, such as double-precision floating-point numbers commonly used in most programs. This is implemented using the [GNU Multiple Precision (GMP) library](https://gmplib.org/) which allows floating-point operations with customizable precision. To maximize performance, the evaluation of the Padé model uses the [Wallis algorithm](https://numerical.recipes/book.html), which minimizes the number of divisions, an operation that is computationally expensive, especially for complex floating-point numbers and even more so for higher-precision complex numbers.
```
But  it is not clear to me that:
- Are both strategies required to stablize the computation, or either of them is enough?
- Why does the 64bit tests pass whereas the 128 bit tests fail when the precision is higher?

Best regards,
Xu He

Below is the test result and the log file is attached. I compiled with gfortran 11.4/ubuntu 22.04/intel 13900K.

  Start 1: test_gx_analytic_continuation_no-greedy_64bit
1/7 Test #1: test_gx_analytic_continuation_no-greedy_64bit ....   Passed    0.21 sec
    Start 2: test_gx_analytic_continuation_greedy_64bit
2/7 Test #2: test_gx_analytic_continuation_greedy_64bit .......   Passed    0.21 sec
    Start 3: test_gx_analytic_continuation_no-greedy_128bit
3/7 Test #3: test_gx_analytic_continuation_no-greedy_128bit ...***Failed    0.22 sec
    Start 4: test_gx_analytic_continuation_greedy_128bit
4/7 Test #4: test_gx_analytic_continuation_greedy_128bit ......***Failed    0.22 sec
    Start 5: test_gx_analytic_continuation_symmetry_64bit
5/7 Test #5: test_gx_analytic_continuation_symmetry_64bit .....   Passed    0.20 sec
    Start 6: test_gx_analytic_continuation_symmetry_128bit
6/7 Test #6: test_gx_analytic_continuation_symmetry_128bit ....***Failed    0.21 sec
    Start 7: test_gx_analytic_continuation_reference_0
7/7 Test #7: test_gx_analytic_continuation_reference_0 ........   Passed    0.20 sec

57% tests passed, 3 tests failed out of 7

Total Test time (real) =   1.46 sec

The following tests FAILED:
	 3 - test_gx_analytic_continuation_no-greedy_128bit (Failed)
	 4 - test_gx_analytic_continuation_greedy_128bit (Failed)
	 6 - test_gx_analytic_continuation_symmetry_128bit (Failed)

[LastTestsFailed.log](https://github.com/user-attachments/files/19683689/LastTestsFailed.log)
