### 2025-04-10 #7859 @mailhexu
Hello,
@HaoZeke @moritzleucke
I have read the paper, tested the code. I think the paper is well written and the code compiles and runs smoothly. The documentation is also quite well written.  I have only one quesion I posted in this issue:
https://github.com/nomad-coe/greenX/issues/156
Best regards,
Xu He
