### 2025-04-10 nomad-coe/greenX#156 @moritzleucke
Hello @mailhexu

Thanks a lot for pointing out this Issue!

**Regarding your second question:** The 128-bit tests fail because they depend on the GMP library, which handles all >64-bit floats in the AC component. When numerical precision >64-bit is requested, but GMP is not linked, a runtime error occurs. This was indeed a bug, as these tests should not be enabled if GMP isn't linked. Based on your feedback, we have updated the testing behavior in pull request #157. Now, if GMP is linked, all regression tests will run, and if GMP is not linked, only the tests that don't depend on GMP (_64bit) will run.

**Regarding your first question:** Using both stabilization strategies is not necessarily required. On the [documentation webpage](https://nomad-coe.github.io/greenX/gx_ac.html) we provide some benchmark results on the accuracy/stability of both strategies. However, the optimal setting depends also on the fitted function. Users should experiment with different combinations (e.g., no greedy + double precision, greedy + double precision, no greedy + multiple precision, greedy + multiple precision) to determine the best configuration for their specific function.

If you have any further questions, feel free to ask!
