### 2025-04-11 #7859 @mailhexu
Hello,
@HaoZeke @moritzleucke
Thanks @moritzleucke for the answer to this question ( https://github.com/nomad-coe/greenX/issues/156 ) and the quick fixing! (https://github.com/nomad-coe/greenX/pull/157)
I recommend the publishing of the paper!
Best regards,
Xu He
