### 2025-04-28 #7859 @moritzleucke
Dear @HaoZeke

Thanks a lot for the for initiating the post-review process.  Furthermore, we want to inform that we completed the list “Author Tasks After Review is Complete”:

- Double check authors and affiliations (including ORCIDs) — **Done**
- Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper. — **Done**, version tag: v2.2
- Archive the release on Zenodo/figshare/etc and post the DOI here. — **Done**, DOI: 10.5281/zenodo.15195060
- Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper. — **Done**
- Make sure that the license listed for the archive is the same as the software license. — **Done** (Apache 2 license)

Best regards,
Moritz Leucke
