### 2025-04-28 649234b
update published arXiv paper
