+++ JOSS/analyticcontinuation/refs.bib
@@ -381,9 +381,12 @@ note ={PMID: 37566917},
 @article{greedy_pade1,
   title={Numerical continued fraction interpolation},
   author={Celis, O. S.},
-  journal={arXiv preprint arXiv:2109.10529},
+  journal={Ukrainskyi Matematychnyi Zhurnal},
   doi = {10.3842/umzh.v74i4.7349},
-  year={2021}
+  year={2024},
+  pages = {568-80},
+  volume = {76},
+  number = {4},
 }
 
 @article{greedy_pade2,
@@ -519,14 +522,19 @@ numpages = {8}
   doi = {10.1021/acs.jctc.6b00511.s001},
   publisher={ACS Publications}
 }
-
-
-
-
-
-
-
-
+@article{kocklauner2025gw,
+  author = {Kockläuner, J. and Golze, D.},
+  title = {GW Plus Cumulant Approach for Predicting Core-Level Shakeup Satellites in Large Molecules},
+  journal = {Journal of Chemical Theory and Computation},
+  volume = {21},
+  number = {6},
+  pages = {3101-3119},
+  year = {2025},
+  doi = {10.1021/acs.jctc.4c01754},
+  note = {PMID: 40029694},
+  URL = {https://doi.org/10.1021/acs.jctc.4c01754},
+  eprint = {https://doi.org/10.1021/acs.jctc.4c01754}
+}
 
 @misc{JohanSchott,
   author = {JohanSchott},
