+++ JOSS/analyticcontinuation/refs.bib
@@ -531,7 +531,6 @@ numpages = {8}
   pages = {3101-3119},
   year = {2025},
   doi = {10.1021/acs.jctc.4c01754},
-  note = {PMID: 40029694},
   URL = {https://doi.org/10.1021/acs.jctc.4c01754},
   eprint = {https://doi.org/10.1021/acs.jctc.4c01754}
 }
