+++ .github/workflows/joss-TimeFrequency-pdf.yml
@@ -6,7 +6,7 @@ jobs:
     name: TimeFrequency Paper
     steps:
       - name: Checkout
-        uses: actions/checkout@v3
+        uses: actions/checkout@v4
       - name: Build draft PDF
         uses: openjournals/openjournals-draft-action@master
         with:
