+++ .github/workflows/action.yml
@@ -26,7 +26,7 @@ jobs:
 
       - uses: actions/setup-python@v1
         with:
-          python-version: '3.7'
+          python-version: '3.8'
 
       - name: Install Cmake
         run: pip3 install cmake
