+++ JOSS/analyticcontinuation/paper.md
@@ -35,9 +35,9 @@ In this work, we present the analytic continuation component of the GreenX libra
 
 # Statement of need
 
-Analytic continuation (AC) is used in different scientific fields where complex analysis is relevant, like mathematical function theory, engineering and theoretical physics/chemistry in e.g.  quantum mechanics [@golze2019gw], quantum field theory [@nekrasov2024analytic], numerical methods for solving differential equations [@lope2002analytic] and real-time propagation methods [@li2020real]. In the following, we discuss the four examples depicted in Figure&#160;1. The first example, shown in the top left of Figure&#160;1, involves the application of AC to model functions, which may include Gamma functions [@luke1975error], Zeta functions [@iriguchi2007estimation], and others. 
+Analytic continuation (AC) is used in different scientific fields where complex analysis is relevant, like mathematical function theory and engineering, for example in numerical methods for solving differential equations [@lope2002analytic]. Another prominent application area is theoretical physics/chemistry in e.g. quantum mechanics [@golze2019gw], quantum field theory [@nekrasov2024analytic] or real-time propagation methods [@li2020real]. In the following, we discuss the four examples depicted in Figure&#160;1. The first example, shown in the top left of Figure&#160;1, involves the application of AC to model functions, which may include Gamma functions [@luke1975error], Zeta functions [@iriguchi2007estimation], and others. 
 
-In quantum field theory, AC can be applied to the frequently arising, complex-valued Green's functions, like the Green's function of the Hubbard model [@schott2016analytic]. However, Green's functions also appear in ab-initio many-body perturbation theory methods like the $GW$ approximation. The $GW$ method [@hedin1965new] is considered the method of choice for predicting band structures of solids as well as electron removal and addition energies of molecules, as measured in direct and indirect photoemission experiments [@golze2019gw]. The complex-valued self energy is a central quantity in the $GW$ method, computed as the convolution of the Green's function $G$ and the screened Coulomb interaction $W$. AC is a frequently used tool for continuing the self energy from the imaginary to the real frequency axis in conventional scaling $GW$ implementations [@gonze2009abinit; @ren2012resolution; @van2015gw; @wilhelm2017periodic] and low-scaling implementations [@liu2016cubic; @wilhelm2018toward;  @forster2020low;@wilhelm2021low; @forster2021low; @forster2021gw100; @forster2023two; @graml2024low]. More recently, AC has also been applied to the screened Coulomb interaction [@springer1998first; @friedrich2019tetrahedron; @duchemin2020robust; @voora2020molecular; @duchemin2021cubic; @samal2022modeling; @cdwac; @kehry2023robust; @kocklauner2025gw] to e.g. reduce the computational scaling associated with core-level excitations [@cdwac]. The AC of a self energy $\Sigma$ and a screened Coulomb interaction $W$ are depicted as the second and third example in the bottom panel of Figure&#160;1.
[…]
