### 2025-05-09 #7859 @dgolze
Dear @HaoZeke,

I believe @moritzleucke has completed the author-side of the post-review checklist. Is there anything else we need to do, or any way we can assist? It would be great to get this over the finish line now.

Thanks again for taking on this JOSS contribution as editor -- it's much appreciated.

Dorothea Golze
