### 2025-05-09 #7859 @moritzleucke
Dear @HaoZeke,

One quick note about the references `@editorialbot check references`:
There has been a [discussion about the missing/skipped DOIs in the old thread](https://github.com/openjournals/joss-reviews/issues/7464#issuecomment-2517841757).

Thanks and best regards,
Moritz Leucke
