### 2025-05-11 736b21d
Add missing GMP targets when find_package(greenX) is called
