### 2025-05-19 #7859 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
