### 2025-05-20 #7859 @moritzleucke
Hey @openjournals/bcm-eics seems that I cannot run the accept command. Is there anything from our side that we can assist with at this point?
