### 2025-05-23 7893bb0
chore: update README.md with new JOSS publication/
I have added the DOI and GitHub badge of the new paper.
